Updated code-prober library and scripts
Also ensure that we report syntax errors.  Currently piggy-backing on
`Program.reports()' to avoid creating more automatic probes that
pollute the UI.

Disabling individual UI elements is not supported by the current
code-prober yet, so this is part is currently non-functional.
Students need to be careful to not accidentally toggle options that
can lead to subtle changes in behaviour, so we should ideally fix that
before the course starts-- however, the "forced defaults" offer a
workaround, by essentially overriding attempts to change configuration
elements.